Added Flag for disabling log. Deleted some warnings.

+++ src/utils/Position.cpp
@@ -19,12 +19,12 @@ Row Position::getRow() const
 	row.setFieldData(fieldData);
 
 	row.addRegisterPerValue<int>(getID());
-	row.addRegisterPerValue<double>(this->m_pos.x);
-	row.addRegisterPerValue<double>(this->m_pos.y);
-	row.addRegisterPerValue<double>(this->m_pos.z);
-	row.addRegisterPerValue<double>(this->m_rot.x);
-	row.addRegisterPerValue<double>(this->m_rot.y);
-	row.addRegisterPerValue<double>(this->m_rot.z);
+	row.addRegisterPerValue<double>((double) this->m_pos.x);
+	row.addRegisterPerValue<double>((double) this->m_pos.y);
+	row.addRegisterPerValue<double>((double) this->m_pos.z);
+	row.addRegisterPerValue<double>((double) this->m_rot.x);
+	row.addRegisterPerValue<double>((double) this->m_rot.y);
+	row.addRegisterPerValue<double>((double) this->m_rot.z);
 
 	return row;
 }
@@ -32,8 +32,8 @@ Row Position::getRow() const
 void Position::operator=(const Row& row)
 {
 	ObjectDB::operator=(row);
-	this->m_pos = Vectorf3D(row.get<double>("pos_x"), row.get<double>("pos_y"), row.get<double>("pos_z"));
-	this->m_rot = Vectorf3D(row.get<double>("rot_roll"), row.get<double>("rot_pitch"), row.get<double>("rot_yaw"));
+	this->m_pos = Vectorf3D((float) row.get<double>("pos_x"), (float) row.get<double>("pos_y"), (float) row.get<double>("pos_z"));
+	this->m_rot = Vectorf3D((float) row.get<double>("rot_roll"), (float) row.get<double>("rot_pitch"), (float) row.get<double>("rot_yaw"));
 }
 
 void to_json(json& j, const Vectorf3D& m)
@@ -43,9 +43,9 @@ void to_json(json& j, const Vectorf3D& m)
 
 void from_json(const json& j, Vectorf3D& m)
 {
-	m.x = j.at("x").get<double>();
-	m.y = j.at("y").get<double>();
-	m.z = j.at("z").get<double>();
+	m.x = j.at("x").get<float>();
+	m.y = j.at("y").get<float>();
+	m.z = j.at("z").get<float>();
 }
 
 void to_json(json& j, const Position& m)
